backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
aiohttp==3.9.1
python-multipart==0.0.6
googlemaps==4.10.0
openai==1.40.0
